requirements.txt
pytest==9.0.2
selene==2.0.0rc9
selenium==4.29.0
allure-pytest==2.15.3
python-dotenv